Flight Code Implementation
See ThirdPersonController.cs for implementation.

Holding Ctrl lets the player in a "fly" state where vertical velocity is set to 0.
If the player uses the JUMP input, he flies upwards.
If the player uses the SPRINT input, he flies downwards.
The player stops flying when they let go of Ctrl.