Flight Control Polishing
When flight is active, the vertical velocity will gradually (but quickly) reach zero instead of a sharp stop.

Organized placement of flight values under appropriate headers.

Notably, I noticed there is a bug that when the vertical velocity gets too high, gravity does NOT affect the player (unless of course the user is in flight mode in which case vertical velocity will fix itself accordingly). I suspect it has to do with however terminal velocity logic was implemented, but am uncertain.